refactor: remove dead typealiases and split expense tag mappers
Remove ModelDebtorAccount, ModelCreditorAccount and ModelExpenseTag
typealiases that had no usages left after the data layer migration.
Move the expense tag mappers out of TransactionMappers.kt into a
dedicated ExpenseTagMappers.kt so each mapper file matches its domain.
