+++ composeApp/src/commonMain/kotlin/com/banko/app/ui/screens/details/DetailsScreenViewModel.kt
@@ -2,12 +2,12 @@ package com.banko.app.ui.screens.details
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.banko.app.ApiExpenseTagRepository
-import com.banko.app.database.Entities.toModelItem
+import com.banko.app.data.repository.ExpenseTagRepository
+import com.banko.app.data.repository.TransactionRepository
 import com.banko.app.domain.AssignExpenseTagToTransactionUseCase
 import com.banko.app.domain.GetAllExpenseTagUseCase
 import com.banko.app.domain.SaveNoteUseCase
-import com.banko.app.data.repository.TransactionRepository
+import com.banko.app.ui.models.toUi
 import com.banko.app.ui.utils.ErrorState
 import com.banko.app.ui.utils.classifyError
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -16,7 +16,7 @@ import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 
 class DetailsScreenViewModel(
-    private val apiTagRepository: ApiExpenseTagRepository,
+    private val expenseTagRepository: ExpenseTagRepository,
     private val updateTransactionUseCase: AssignExpenseTagToTransactionUseCase,
     private val getExpenseTags: GetAllExpenseTagUseCase,
     private val saveNoteUseCase: SaveNoteUseCase,
@@ -32,7 +32,7 @@ class DetailsScreenViewModel(
     fun getExpenseTags() {
         viewModelScope.launch {
             getExpenseTags.invoke().collect { result ->
-                _screenState.update { it.copy(expenseTags = result.mapNotNull { it?.toModelItem() }) }
+                _screenState.update { it.copy(expenseTags = result.map { it.toUi() }) }
             }
         }
     }
@@ -41,7 +41,7 @@ class DetailsScreenViewModel(
         viewModelScope.launch {
             val previousTagId = transactionRepository.getTransactionById(id)?.expenseTag?.id
             try {
-                apiTagRepository.assignExpenseTag(id, expenseTagId)
+                expenseTagRepository.assignExpenseTag(id, expenseTagId)
                 updateTransactionUseCase.invoke(transactionId = id, expenseTagId = expenseTagId)
                 _screenState.update { it.copy(error = null) }
             } catch (e: Exception) {
